+++ cpp/open3d/geometry/TriangleMesh.cpp
@@ -160,280 +160,6 @@ TriangleMesh &TriangleMesh::ComputeAdjacencyList() {
     return *this;
 }
 
-std::shared_ptr<TriangleMesh> TriangleMesh::FilterSharpen(
-        int number_of_iterations, double strength, FilterScope scope) const {
-    bool filter_vertex =
-            scope == FilterScope::All || scope == FilterScope::Vertex;
-    bool filter_normal =
-            (scope == FilterScope::All || scope == FilterScope::Normal) &&
-            HasVertexNormals();
-    bool filter_color =
-            (scope == FilterScope::All || scope == FilterScope::Color) &&
-            HasVertexColors();
-
-    std::vector<Eigen::Vector3d> prev_vertices = vertices_;
-    std::vector<Eigen::Vector3d> prev_vertex_normals = vertex_normals_;
-    std::vector<Eigen::Vector3d> prev_vertex_colors = vertex_colors_;
-
-    std::shared_ptr<TriangleMesh> mesh = std::make_shared<TriangleMesh>();
-    mesh->vertices_.resize(vertices_.size());
-    mesh->vertex_normals_.resize(vertex_normals_.size());
-    mesh->vertex_colors_.resize(vertex_colors_.size());
-    mesh->triangles_ = triangles_;
-    mesh->adjacency_list_ = adjacency_list_;
-    if (!mesh->HasAdjacencyList()) {
-        mesh->ComputeAdjacencyList();
-    }
-
-    for (int iter = 0; iter < number_of_iterations; ++iter) {
-        for (size_t vidx = 0; vidx < mesh->vertices_.size(); ++vidx) {
-            Eigen::Vector3d vertex_sum(0, 0, 0);
-            Eigen::Vector3d normal_sum(0, 0, 0);
-            Eigen::Vector3d color_sum(0, 0, 0);
-            for (int nbidx : mesh->adjacency_list_[vidx]) {
-                if (filter_vertex) {
-                    vertex_sum += prev_vertices[nbidx];
-                }
-                if (filter_normal) {
-                    normal_sum += prev_vertex_normals[nbidx];
-                }
-                if (filter_color) {
-                    color_sum += prev_vertex_colors[nbidx];
-                }
-            }
-
-            size_t nb_size = mesh->adjacency_list_[vidx].size();
-            if (filter_vertex) {
-                mesh->vertices_[vidx] =
-                        prev_vertices[vidx] +
-                        strength * (prev_vertices[vidx] * nb_size - vertex_sum);
-            }
-            if (filter_normal) {
-                mesh->vertex_normals_[vidx] =
-                        prev_vertex_normals[vidx] +
-                        strength * (prev_vertex_normals[vidx] * nb_size -
-                                    normal_sum);
-            }
-            if (filter_color) {
-                mesh->vertex_colors_[vidx] =
-                        prev_vertex_colors[vidx] +
-                        strength * (prev_vertex_colors[vidx] * nb_size -
-                                    color_sum);
-            }
-        }
-        if (iter < number_of_iterations - 1) {
-            std::swap(mesh->vertices_, prev_vertices);
-            std::swap(mesh->vertex_normals_, prev_vertex_normals);
-            std::swap(mesh->vertex_colors_, prev_vertex_colors);
-        }
-    }
-
-    return mesh;
-}
-
-std::shared_ptr<TriangleMesh> TriangleMesh::FilterSmoothSimple(
-        int number_of_iterations, FilterScope scope) const {
-    bool filter_vertex =
-            scope == FilterScope::All || scope == FilterScope::Vertex;
-    bool filter_normal =
-            (scope == FilterScope::All || scope == FilterScope::Normal) &&
-            HasVertexNormals();
-    bool filter_color =
-            (scope == FilterScope::All || scope == FilterScope::Color) &&
-            HasVertexColors();
-
-    std::vector<Eigen::Vector3d> prev_vertices = vertices_;
-    std::vector<Eigen::Vector3d> prev_vertex_normals = vertex_normals_;
-    std::vector<Eigen::Vector3d> prev_vertex_colors = vertex_colors_;
-
-    std::shared_ptr<TriangleMesh> mesh = std::make_shared<TriangleMesh>();
-    mesh->vertices_.resize(vertices_.size());
-    mesh->vertex_normals_.resize(vertex_normals_.size());
-    mesh->vertex_colors_.resize(vertex_colors_.size());
-    mesh->triangles_ = triangles_;
-    mesh->adjacency_list_ = adjacency_list_;
-    if (!mesh->HasAdjacencyList()) {
-        mesh->ComputeAdjacencyList();
-    }
-
-    for (int iter = 0; iter < number_of_iterations; ++iter) {
-        for (size_t vidx = 0; vidx < mesh->vertices_.size(); ++vidx) {
-            Eigen::Vector3d vertex_sum(0, 0, 0);
-            Eigen::Vector3d normal_sum(0, 0, 0);
-            Eigen::Vector3d color_sum(0, 0, 0);
-            for (int nbidx : mesh->adjacency_list_[vidx]) {
-                if (filter_vertex) {
-                    vertex_sum += prev_vertices[nbidx];
-                }
-                if (filter_normal) {
-                    normal_sum += prev_vertex_normals[nbidx];
-                }
-                if (filter_color) {
-                    color_sum += prev_vertex_colors[nbidx];
-                }
-            }
-
-            size_t nb_size = mesh->adjacency_list_[vidx].size();
-            if (filter_vertex) {
-                mesh->vertices_[vidx] =
-                        (prev_vertices[vidx] + vertex_sum) / (1 + nb_size);
-            }
-            if (filter_normal) {
-                mesh->vertex_normals_[vidx] =
-                        (prev_vertex_normals[vidx] + normal_sum) /
-                        (1 + nb_size);
-            }
-            if (filter_color) {
-                mesh->vertex_colors_[vidx] =
-                        (prev_vertex_colors[vidx] + color_sum) / (1 + nb_size);
-            }
-        }
-        if (iter < number_of_iterations - 1) {
-            std::swap(mesh->vertices_, prev_vertices);
-            std::swap(mesh->vertex_normals_, prev_vertex_normals);
-            std::swap(mesh->vertex_colors_, prev_vertex_colors);
-        }
-    }
-    return mesh;
-}
-
-void TriangleMesh::FilterSmoothLaplacianHelper(
-        std::shared_ptr<TriangleMesh> &mesh,
-        const std::vector<Eigen::Vector3d> &prev_vertices,
-        const std::vector<Eigen::Vector3d> &prev_vertex_normals,
-        const std::vector<Eigen::Vector3d> &prev_vertex_colors,
-        const std::vector<std::unordered_set<int>> &adjacency_list,
-        double lambda,
-        bool filter_vertex,
-        bool filter_normal,
-        bool filter_color) const {
-    for (size_t vidx = 0; vidx < mesh->vertices_.size(); ++vidx) {
-        Eigen::Vector3d vertex_sum(0, 0, 0);
-        Eigen::Vector3d normal_sum(0, 0, 0);
-        Eigen::Vector3d color_sum(0, 0, 0);
-        double total_weight = 0;
-        for (int nbidx : mesh->adjacency_list_[vidx]) {
-            auto diff = prev_vertices[vidx] - prev_vertices[nbidx];
-            double dist = diff.norm();
-            double weight = 1. / (dist + 1e-12);
-            total_weight += weight;
-
-            if (filter_vertex) {
-                vertex_sum += weight * prev_vertices[nbidx];
-            }
-            if (filter_normal) {
-                normal_sum += weight * prev_vertex_normals[nbidx];
-            }
-            if (filter_color) {
-                color_sum += weight * prev_vertex_colors[nbidx];
-            }
-        }
-
-        if (filter_vertex) {
-            mesh->vertices_[vidx] =
-                    prev_vertices[vidx] +
-                    lambda * (vertex_sum / total_weight - prev_vertices[vidx]);
-        }
-        if (filter_normal) {
-            mesh->vertex_normals_[vidx] = prev_vertex_normals[vidx] +
-                                          lambda * (normal_sum / total_weight -
-                                                    prev_vertex_normals[vidx]);
-        }
-        if (filter_color) {
-            mesh->vertex_colors_[vidx] = prev_vertex_colors[vidx] +
-                                         lambda * (color_sum / total_weight -
-                                                   prev_vertex_colors[vidx]);
-        }
-    }
-}
-
-std::shared_ptr<TriangleMesh> TriangleMesh::FilterSmoothLaplacian(
-        int number_of_iterations, double lambda, FilterScope scope) const {
-    bool filter_vertex =
-            scope == FilterScope::All || scope == FilterScope::Vertex;
-    bool filter_normal =
-            (scope == FilterScope::All || scope == FilterScope::Normal) &&
-            HasVertexNormals();
-    bool filter_color =
-            (scope == FilterScope::All || scope == FilterScope::Color) &&
-            HasVertexColors();
-
-    std::vector<Eigen::Vector3d> prev_vertices = vertices_;
-    std::vector<Eigen::Vector3d> prev_vertex_normals = vertex_normals_;
-    std::vector<Eigen::Vector3d> prev_vertex_colors = vertex_colors_;
-
-    std::shared_ptr<TriangleMesh> mesh = std::make_shared<TriangleMesh>();
-    mesh->vertices_.resize(vertices_.size());
-    mesh->vertex_normals_.resize(vertex_normals_.size());
-    mesh->vertex_colors_.resize(vertex_colors_.size());
-    mesh->triangles_ = triangles_;
-    mesh->adjacency_list_ = adjacency_list_;
-    if (!mesh->HasAdjacencyList()) {
-        mesh->ComputeAdjacencyList();
-    }
-
-    for (int iter = 0; iter < number_of_iterations; ++iter) {
-        FilterSmoothLaplacianHelper(mesh, prev_vertices, prev_vertex_normals,
-                                    prev_vertex_colors, mesh->adjacency_list_,
-                                    lambda, filter_vertex, filter_normal,
-                                    filter_color);
-        if (iter < number_of_iterations - 1) {
-            std::swap(mesh->vertices_, prev_vertices);
-            std::swap(mesh->vertex_normals_, prev_vertex_normals);
-            std::swap(mesh->vertex_colors_, prev_vertex_colors);
-        }
-    }
-    return mesh;
-}
-
-std::shared_ptr<TriangleMesh> TriangleMesh::FilterSmoothTaubin(
-        int number_of_iterations,
-        double lambda,
-        double mu,
-        FilterScope scope) const {
-    bool filter_vertex =
-            scope == FilterScope::All || scope == FilterScope::Vertex;
-    bool filter_normal =
-            (scope == FilterScope::All || scope == FilterScope::Normal) &&
-            HasVertexNormals();
-    bool filter_color =
-            (scope == FilterScope::All || scope == FilterScope::Color) &&
-            HasVertexColors();
-
-    std::vector<Eigen::Vector3d> prev_vertices = vertices_;
-    std::vector<Eigen::Vector3d> prev_vertex_normals = vertex_normals_;
-    std::vector<Eigen::Vector3d> prev_vertex_colors = vertex_colors_;
-
-    std::shared_ptr<TriangleMesh> mesh = std::make_shared<TriangleMesh>();
-    mesh->vertices_.resize(vertices_.size());
-    mesh->vertex_normals_.resize(vertex_normals_.size());
-    mesh->vertex_colors_.resize(vertex_colors_.size());
-    mesh->triangles_ = triangles_;
-    mesh->adjacency_list_ = adjacency_list_;
-    if (!mesh->HasAdjacencyList()) {
-        mesh->ComputeAdjacencyList();
-    }
-    for (int iter = 0; iter < number_of_iterations; ++iter) {
-        FilterSmoothLaplacianHelper(mesh, prev_vertices, prev_vertex_normals,
-                                    prev_vertex_colors, mesh->adjacency_list_,
-                                    lambda, filter_vertex, filter_normal,
-                                    filter_color);
-        std::swap(mesh->vertices_, prev_vertices);
-        std::swap(mesh->vertex_normals_, prev_vertex_normals);
-        std::swap(mesh->vertex_colors_, prev_vertex_colors);
-        FilterSmoothLaplacianHelper(mesh, prev_vertices, prev_vertex_normals,
-                                    prev_vertex_colors, mesh->adjacency_list_,
-                                    mu, filter_vertex, filter_normal,
-                                    filter_color);
-        if (iter < number_of_iterations - 1) {
-            std::swap(mesh->vertices_, prev_vertices);
-            std::swap(mesh->vertex_normals_, prev_vertex_normals);
-            std::swap(mesh->vertex_colors_, prev_vertex_colors);
-        }
-    }
-    return mesh;
-}
-
 TriangleMesh &TriangleMesh::RemoveDuplicatedVertices() {
     typedef std::tuple<double, double, double> Coordinate3;
     std::unordered_map<Coordinate3, size_t,
@@ -1394,39 +1120,6 @@ std::shared_ptr<TriangleMesh> TriangleMesh::Crop(
         return std::make_shared<TriangleMesh>();
     }
     return SelectByIndex(bbox.GetPointIndicesWithinBoundingBox(vertices_));
-}
-
-std::unordered_map<Eigen::Vector2i,
-                   double,
-                   utility::hash_eigen::hash<Eigen::Vector2i>>
-TriangleMesh::ComputeEdgeWeightsCot(
-        const std::unordered_map<Eigen::Vector2i,
-                                 std::vector<int>,
-                                 utility::hash_eigen::hash<Eigen::Vector2i>>
-                &edges_to_vertices,
-        double min_weight) const {
-    std::unordered_map<Eigen::Vector2i, double,
-                       utility::hash_eigen::hash<Eigen::Vector2i>>
-            weights;
-    for (const auto &edge_v2s : edges_to_vertices) {
-        Eigen::Vector2i edge = edge_v2s.first;
-        double weight_sum = 0;
-        int N = 0;
-        for (int v2 : edge_v2s.second) {
-            Eigen::Vector3d a = vertices_[edge(0)] - vertices_[v2];
-            Eigen::Vector3d b = vertices_[edge(1)] - vertices_[v2];
-            double weight = a.dot(b) / (a.cross(b)).norm();
-            weight_sum += weight;
-            N++;
-        }
-        double weight = N > 0 ? weight_sum / N : 0;
-        if (weight < min_weight) {
-            weights[edge] = min_weight;
-        } else {
-            weights[edge] = weight;
-        }
-    }
-    return weights;
 }
 
 }  // namespace geometry
